Fix memory corruption, DTL write recursion, and error-handling gaps across gateway/s7shell/scl
Replaces the heavyweight class-based Error{code,message}/.string() error
type with a lightweight enum-based SclError/S7Error + free toString()
(fmt-formattable) across scl, gateway, and s7shell, and introduces a
dedicated ShellError type (sgrn/s7shell/errors.hpp) with S7Error/SclError
bridging and throwScriptException() for consistent AngelScript exception
propagation. Also gives DbIOProvider its own error type (DbIoError),
finishing the migration off PlcMemoryError that was left mid-refactor and
failing to build. Fixes several correctness and safety bugs uncovered
along the way:

* SchemaVM: fix stack buffer over-read/memory corruption in the generic
  int64 field setter — a fixed uint8_t tmp[16] was encoded via
  encodeScalar() and then copied out using typeSpanBytes(type, count),
  which can exceed 16 bytes for arrays/wide fields, corrupting adjacent
  DB memory on write. Buffer is now sized dynamically from
  typeSpanBytes(...).value_or(...) and passed consistently to both
  encodeScalar() and writeFieldToMemory().

* OPC UA adapter: fix infinite recursive write loop on DTL/LDT (temporal)
  fields. Server-initiated writes (event push) re-entered open62541's
  DataSource write callback, which wrote the value back to PLC memory and
  re-emitted the event; since UA's 100ns precision never round-trips
  bit-exact against the PLC's 1ns representation, the change-detection
  memcmp never converged, corrupting the field under continuous rewrite.
  Guarded with a thread_local g_is_internal_opcua_write flag so
  self-triggered writes short-circuit instead of looping.

* s7codec: single-quoted char literals ('A') in shell/JSON value parsing
  now decode to their byte value instead of being silently rejected,
  fixing CHAR/BYTE field assignment from the shell.

* SchemaVM string setter: fix STRING[N] encoding using the wrong capacity
  (count==1 for a scalar string field) instead of the schema's
  string_capacity (N), which corrupted/truncated writes for strings
  longer than one character. Also special-cases CHAR/WCHAR so a
  string-typed setter call correctly encodes a single character rather
  than a string blob.

* s7codec::primitiveSize()/typeSpanBytes() now return std::optional
  instead of silently defaulting on unknown/unsupported types; all call
  sites updated to handle the failure case explicitly instead of
  propagating a bogus 0/garbage size into buffer allocation and memcpy
  calls.

* ScriptDataBlock::get()/get(size_t): no longer silently succeed. The
  size-less get() now reports failure (last_op_ok_/last_op_err_) instead
  of returning as if a read had happened when no schema/size is known;
  offline (not-connected) get(size_t) now explicitly flags the result as
  a local-memory fetch rather than pretending a PLC round trip occurred.

* DbIOProvider (gateway/twin): introduces DbIoError, replacing the
  PlcMemoryError type it had been force-fit into (and before that,
  SclError) — DbIOProvider sits at the intersection of S7Error (wire),
  PlcMemoryError (local shadow arena), and SclError (schema/field
  resolution) and needed a domain of its own, with each internal failure
  bridged from wherever it actually originated (fromS7Error,
  fromPlcMemoryError, fromSclError) instead of being collapsed into one
  enum that didn't have members for half the failure modes. get()/put()
  now check t_client.isConnected() up front and fail fast with
  DbIoError::NotConnected instead of only discovering it via a failed
  readMultiVars/writeMultiVars. Downstream: s7shell/errors.hpp gains
  fromDbIoError() -> ShellError; ScriptDataBlock gains a setOpResult()
  overload for Result<T, DbIoError> and its val()/setVal()/get()/
  readScalar() are fixed to bridge through it instead of the wrong
  fromS7Error(); S7BatchEngine.hpp gains a matching proto_bridge()
  overload so its DbIOProvider instantiation (buildGetItems/
  buildPutItems/commitLocalPut) compiles again.

* PlcMemory::writeDbMemory and friends: bounds/null checks normalized
  through SGRN_RETURN_IF*/SGRN_RETURN_IF_NULL macros for consistency.

Also included in this pass:
* scl/errors.hpp: new SclError enum (adds BufferOverflow,
  RecursionLimitExceeded, CodecFailure, InvalidFormat, etc.) with
  toString() and an fmt::formatter specialization.
* DTL: register an AngelScript factory (asBEHAVE_FACTORY) so `DTL myVar;`
  initializes to the current sim-clock time, plus dtlToTimeMs()/
  dtlToTodMs() conversion helpers.
* Mechanical rename fallout from the Error -> SclError/S7Error migration
  across ~90 files (gateway adapters/wrappers/twin, s7shell facades and
  bindings, scl schema/codec), including OPC UA, EtherNet/IP, Modbus,
  and HTTP adapter error paths.
* documentation/gateway/config.md and simulation configs updated with
  non-placeholder example tokens.

No functional change intended outside the fixes listed above; the
Error -> SclError/S7Error/DbIoError rename is mechanical.

+++ sgrn/scl/include/sgrn/scl/types/DbField.hpp
@@ -11,17 +11,17 @@ namespace sgrn::scl
 {
 struct DbField {
     std::string name;
-    int offset{0};
-    int bit_index{0};
+    int32_t offset{0};
+    uint8_t bit_index{0};
     DataType type{DataType::Byte};
-    int count{0};
+    uint32_t count{0};
     int array_lower_bound{0};
     int array_upper_bound{0};
-    int string_capacity{0};
+    uint32_t string_capacity{0};
 
     std::string udt_name;
     std::vector<DbField> children;
-    int struct_size{0};
+    uint32_t struct_size{0};
     s7codec::Endian endianness{s7codec::Endian::Big};
     std::optional<std::string> unit;
     std::optional<double> min_val;
